# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Gym Project API
  version: 1.0.0
  description: JSON REST API interacting with the gym project backend
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://your-production-url.com/api
    description: Production server
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: auth_key
      description: Authentication key obtained from /auth/login
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: array
          items:
            type: object
...
    Booking:
      type: object
      properties:
        booking_id:
          type: integer
        user_id:
          type: integer
        session_id:
          type: integer
        booking_date:
          type: string
          format: date
        status:
          type: string
    CreateBookingRequest:
      type: object
      required:
        - sessionId
      properties:
        sessionId:
          type: integer
          example: 42
...
paths:
...
  /bookings:
    get:
      summary: Get bookings
      tags:
        - Bookings
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Booking'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Access denied
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500':
          description: Internal server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

    post:
      summary: Book a session
      tags:
        - Bookings
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
      responses:
        '201':
          description: Session booked successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Members only
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Members only
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500':
          description: Internal server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /bookings/{id}:
    delete:
      summary: Cancel a booking
      tags:
        - Bookings
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Booking cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Access denied
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Booking not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500':
          description: Internal server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  # ==================== XML EXPORT ROUTES ====================
...
  /xml/bookings:
    get:
      summary: Export all bookings as XML
      tags:
        - Export
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Bookings exported as XML
          content:
            application/xml:
              schema:
                type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Access denied
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500':
          description: Internal server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
security:
  - ApiKeyAuth: []